Convert shim/media to typescript
