+++ ext/src/shim/media/classes/EditTracksInfoRequest.ts
@@ -1,0 +1,10 @@
+"use strict";
+
+export default class EditTracksInfoRequest {
+    public requestId = 0;
+
+    constructor (
+            public activeTrackIds: number[] = null
+          , public textTrackStyle: string = null) {
+    }
+}
