--- ext/src/shim/media/classes/GetStatusRequest.js
@@ -1,7 +1,0 @@
-"use strict";
-
-export default class GetStatusRequest {
-    constructor () {
-        this.customData = null;
-    }
-}
