+++ ext/src/shim/media/classes/GetStatusRequest.ts
@@ -1,0 +1,5 @@
+"use strict";
+
+export default class GetStatusRequest {
+    public customData: any = null;
+}
