--- ext/src/shim/media/classes/MediaInfo.js
@@ -1,16 +1,0 @@
-"use strict";
-
-import { StreamType } from "../enums";
-
-export default class MediaInfo {
-	constructor (contentId, contentType) {
-		this.contentId = contentId;
-		this.contentType = contentType;
-		this.customData = null;
-		this.duration = null;
-		this.metadata = null;
-		this.streamType = StreamType.BUFFERED;
-		this.textTrackStyle = null;
-		this.tracks = null;
-	}
-}
