--- ext/src/shim/media/classes/PauseRequest.js
@@ -1,7 +1,0 @@
-"use strict";
-
-export default class PauseRequest {
-    constructor () {
-        this.customData = null;
-    }
-}
