+++ ext/src/shim/media/classes/PauseRequest.ts
@@ -1,0 +1,5 @@
+"use strict";
+
+export default class PauseRequest {
+    public customData: any = null;
+}
