--- ext/src/shim/media/classes/PhotoMediaMetadata.js
@@ -1,19 +1,0 @@
-"use strict";
-
-import { MetadataType } from "../enums";
-
-export default class PhotoMediaMetadata {
-  constructor () {
-        this.artist = null;
-        this.creationDateTime = null;
-        this.height = null;
-        this.images = null;
-        this.latitude = null;
-        this.location = null;
-        this.longitude = null;
-        this.metadataType = MetadataType.PHOTO;
-        this.title = null;
-        this.type = MetadataType.PHOTO;
-        this.width = null;
-    }
-}
