--- ext/src/shim/media/classes/PlayRequest.js
@@ -1,7 +1,0 @@
-"use strict";
-
-export default class PlayRequest {
-    constructor () {
-        this.customData = null;
-    }
-}
