+++ ext/src/shim/media/classes/PlayRequest.ts
@@ -1,0 +1,5 @@
+"use strict";
+
+export default class PlayRequest {
+    public customData: any = null;
+}
