--- ext/src/shim/media/classes/QueueRemoveItemsRequest.js
@@ -1,11 +1,0 @@
-"use strict";
-
-export default class QueueRemoveItemsRequest {
-    constructor (itemIdsToRemove) {
-        this.customData = null;
-        this.itemIds = itemIdsToRemove;
-        this.requestId = null;
-        this.sessionId = null;
-        this.type = "QUEUE_REMOVE";
-    }
-}
