+++ ext/src/shim/media/classes/QueueRemoveItemsRequest.ts
@@ -1,0 +1,11 @@
+"use strict";
+
+export default class QueueRemoveItemsRequest {
+    public customData: any = null;
+    public requestId: number = null;
+    public sessionId: string = null;
+    public type: string = "QUEUE_REMOVE";
+
+    constructor (
+            public itemIds: number[]) {}
+}
