+++ ext/src/shim/media/classes/QueueSetPropertiesRequest.ts
@@ -1,0 +1,9 @@
+"use strict";
+
+export default class QueueSetPropertiesRequest {
+    public customData: any = null;
+    public repeatMode: string = null;
+    public requestId: number = null;
+    public sessionId: string = null;
+    public type: string = "QUEUE_UPDATE";
+}
