--- ext/src/shim/media/classes/SeekRequest.js
@@ -1,9 +1,0 @@
-"use strict";
-
-export default class SeekRequest {
-    constructor () {
-        this.currentTime = null;
-        this.customData = null;
-        this.resumeState = null;
-    }
-}
