+++ ext/src/shim/media/classes/SeekRequest.ts
@@ -1,0 +1,7 @@
+"use strict";
+
+export default class SeekRequest {
+    public currentTime: number = null;
+    public customData: any = null;
+    public resumeState: string = null;
+}
