--- ext/src/shim/media/classes/StopRequest.js
@@ -1,7 +1,0 @@
-"use strict";
-
-export default class StopRequest {
-    constructor () {
-        this.customData = null;
-    }
-}
