+++ ext/src/shim/media/classes/StopRequest.ts
@@ -1,0 +1,5 @@
+"use strict";
+
+export default class StopRequest {
+    public customData: any = null;
+}
