--- ext/src/shim/media/classes/TextTrackStyle.js
@@ -1,18 +1,0 @@
-"use strict";
-
-export default class TextTrackStyle {
-    constructor () {
-        this.backgroundColor = null;
-        this.customData = null;
-        this.edgeColor = null;
-        this.edgeType = null;
-        this.fontFamily = null;
-        this.fontGenericFamily = null;
-        this.fontScale = null;
-        this.fontStyle = null;
-        this.foregroundColor = null;
-        this.windowColor = null;
-        this.windowRoundedCornerRadius = null;
-        this.windowType = null;
-    }
-}
