+++ ext/src/shim/media/classes/TextTrackStyle.ts
@@ -1,0 +1,16 @@
+"use strict";
+
+export default class TextTrackStyle {
+    public backgroundColor: string = null;
+    public customData: any = null;
+    public edgeColor: string = null;
+    public edgeType: string = null;
+    public fontFamily: string = null;
+    public fontGenericFamily: string = null;
+    public fontScale: number = null;
+    public fontStyle: string = null;
+    public foregroundColor: string = null;
+    public windowColor: string = null;
+    public windowRoundedCornerRadius: number = null;
+    public windowType: string = null;
+}
