--- ext/src/shim/media/classes/Track.js
@@ -1,14 +1,0 @@
-"use strict";
-
-export default class Track {
-    constructor (trackId, trackType) {
-        this.customData = null;
-        this.language = null;
-        this.name = null;
-        this.subtype = null;
-        this.trackContentId = null;
-        this.trackContentType = null;
-        this.trackId = trackId;
-        this.type = trackType;
-    }
-}
