+++ ext/src/shim/media/classes/Track.ts
@@ -1,0 +1,14 @@
+"use strict";
+
+export default class Track {
+    public customData: any = null;
+    public language: string = null;
+    public name: string = null;
+    public subtype: string = null;
+    public trackContentId: string = null;
+    public trackContentType: string = null;
+
+    constructor (
+            public trackId: number
+          , public type: string) {}
+}
