--- ext/src/shim/media/classes/VolumeRequest.js
@@ -1,8 +1,0 @@
-"use strict";
-
-export default class VolumeRequest {
-    constructor (volume) {
-        this.customData = null;
-        this.volume = volume;
-    }
-}
