+++ ext/src/shim/media/classes/VolumeRequest.ts
@@ -1,0 +1,11 @@
+"use strict";
+
+import Volume from "../../cast/classes/Volume";
+
+
+export default class VolumeRequest {
+    public customData: any = null;
+
+    constructor (
+            public volume: Volume) {}
+}
